Updated color-coding
New features were added to some functions to allow for better colour coding of data from the fluorescence anisotropy module. All changes are detailed below:

The ‘scinot’ parameter was renamed into ‘str_format’ and the ‘visualise’ and ‘labelwell’ functions updated accordingly. A new functionality was implemented into the ‘labelwell’ function allowing for formatting numbers to 1 decimal place (str_format=’percent’). 

An import statement for warnings module was added.

The header dictionary was reversed to the original version containing header names and data types characteristic for the calcium flex module.

A new function ‘get_color_dict’ was created. It took over some of the functionality of the previous ‘wellcolour’ function, that is it generates a dictionary of labels and colours. New feature include generating an array of the ‘non-categorical’ data (e.g. intensity, anisotropy) that has a different value in each well and normalising it linearly or logarithmically to bring the values to range 0 to 1 and getting a colour for each value. This allows the colour of the well to represent the magnitude of the number. The previous functionality, i.e. generating an array of uniformly spaced values between 0 and 1, was retained for ‘categorical’ data (Type, Protein Name, etc.). The new functionality is executed only if the appropriate keyword arguments are passed: ‘categorical’, ‘scale’, ‘blank_yellow’. 

The ‘wellcolour’ function was modified. It accepts a new argument ‘colordict’ which is a dictionary with colours returned by the ‘get_color_dict’ function. The function now only gets a colour from the dictionary and does not generate it. A functionality to assign the colour as yellow for the blank wells was added (controlled by the keyword argument ‘blank_yellow’) and functionality from the ‘wellcolour2’ function was moved to this function (controlled by keyword argument ‘to_plot’).

The ‘get_color_dict’ function was created so that the colour dictionary is not generated (and calculations associated with it performed) every time the well colour function is called. Instead, the colour dictionary is created at the beginning of the ‘vsualise’ and ‘visualise_all_series’ functions. The ‘visualise’ function now accepts keyword arguments that are passed to the ‘wellcolour’ and ‘get_colour_dict’ functions to control new functionality. 

The ‘wellcolour2’ function now only calls the ‘get_colour_dict’ and ‘wellcolour’ functions.