Update plate_map.py
All if-else statements using kwargs were updated to include a check for the presence of desired keyword argument in kwargs.

A new parameter 'header_type' was added to the 'empty_map' functions. It controls whether the platemap df is created with headers characteristic for long or short platemap. The 'plate_map' and 'short_map' functions were modified to include the new parameter in the 'empty_map' function calls.